basic test for localtable
