--- cypress/component/LocalTable.cy.tsx
@@ -1,5 +1,0 @@
-describe('LocalTable.cy.tsx', () => {
-  it('playground', () => {
-    // cy.mount()
-  })
-})
