+++ cypress/support/component.ts
@@ -15,6 +15,7 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands'
+import '../../app/react/App/styles/globals.css';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
@@ -25,15 +26,9 @@ import { mount } from 'cypress/react18'
 // your custom command.
 // Alternatively, can be defined in cypress/support/component.d.ts
 // with a <reference path="./component" /> at the top of your spec.
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      mount: typeof mount
-    }
-  }
-}
 
 Cypress.Commands.add('mount', mount)
+
 
 // Example use:
 // cy.mount(<MyComponent />)
